Some improvements (see description)
- Agents have their own display() and displayError() to indicate which
agent is printing the information.
- 'ElevatorMessage' bug fix.
- Contract net methods calling it superclass methods.
- expectedTimeToFloor() bug fixed, so it uses the data given as program
argument.
- Code refactoring.
- Assuring that the destination floor is different than the initial
floor.
- Using our own functions for 'random'.
- Sending destination floor from 'Building' to 'Elevator' (it was
missing).

+++ ElevatorMessage.java
@@ -13,7 +13,7 @@ public class ElevatorMessage {
         if (matcher.matches()) {
             this.initialFloor = Integer.parseInt(matcher.group(1));
             this.destinationFloor = Integer.parseInt(matcher.group(2));
-            this.timeToInitialFloor = Integer.parseInt(matcher.group(3));
+            this.timeToInitialFloor = Long.parseLong(matcher.group(3));
         } else
             throw new IllegalArgumentException("Invalid format for message: " + message);
     }
